show suggestions when typing in an origin
