+++ pkg/gui/branches_panel.go
@@ -7,8 +7,6 @@ import (
 	"github.com/jesseduffield/lazygit/pkg/commands"
 	"github.com/jesseduffield/lazygit/pkg/commands/models"
 	"github.com/jesseduffield/lazygit/pkg/commands/oscommands"
-	"github.com/jesseduffield/lazygit/pkg/gui/presentation"
-	"github.com/jesseduffield/lazygit/pkg/gui/types"
 	"github.com/jesseduffield/lazygit/pkg/utils"
 )
 
@@ -220,7 +218,7 @@ func (gui *Gui) handleCheckoutRef(ref string, options handleCheckoutRefOptions) 
 func (gui *Gui) handleCheckoutByName() error {
 	return gui.prompt(promptOpts{
 		title:               gui.Tr.BranchName + ":",
-		findSuggestionsFunc: gui.findBranchNameSuggestions,
+		findSuggestionsFunc: gui.getBranchNameSuggestionsFunc(),
 		handleConfirm: func(response string) error {
 			return gui.handleCheckoutRef(response, handleCheckoutRefOptions{
 				span: "Checkout branch",
@@ -533,32 +531,6 @@ func (gui *Gui) handleNewBranchOffCurrentItem() error {
 			return gui.refreshSidePanels(refreshOptions{mode: ASYNC})
 		},
 	})
-}
-
-func (gui *Gui) getBranchNames() []string {
-	result := make([]string, len(gui.State.Branches))
-
-	for i, branch := range gui.State.Branches {
-		result[i] = branch.Name
-	}
-
-	return result
-}
-
-func (gui *Gui) findBranchNameSuggestions(input string) []*types.Suggestion {
-	branchNames := gui.getBranchNames()
-
-	matchingBranchNames := utils.FuzzySearch(sanitizedBranchName(input), branchNames)
-
-	suggestions := make([]*types.Suggestion, len(matchingBranchNames))
-	for i, branchName := range matchingBranchNames {
-		suggestions[i] = &types.Suggestion{
-			Value: branchName,
-			Label: presentation.GetBranchTextStyle(branchName).Sprint(branchName),
-		}
-	}
-
-	return suggestions
 }
 
 // sanitizedBranchName will remove all spaces in favor of a dash "-" to meet
